Organizing RecipeIngredientRepository
